+++ search/views.py
@@ -1,3 +1,4 @@
+from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
@@ -5,12 +6,18 @@ from rest_framework.views import APIView
 
 from quizzes.serializers import LessonSerializer, QuizSerializer
 from search.utils import search_lessons, search_quizzes
+from search.schemas import SearchSchema
 
 
 class SearchView(APIView):
     permission_classes = (AllowAny,)
 
-    def get(self, request, *args, **kwargs):  # TODO: schema for get method and query params
+    @swagger_auto_schema(
+        operation_summary="Search by whole content",
+        operation_description="Search by lessons and quizzes",
+        responses={200: SearchSchema}
+    )
+    def get(self, request, *args, **kwargs):
         q = request.GET.get('q', '')
         data = {
             'lessons': LessonSerializer(search_lessons(q).to_queryset(), many=True).data,