+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added lessons model
 * Added questions action
+* Added elasticsearch
 
 ## v0.4.0 - 17.03.2021